+++ saw-remote-api/CHANGELOG.md
@@ -35,6 +35,7 @@
 * Add `"slice"` and `"slice range"` `setup value`s representing slices in MIR
   verification. Attempting to use these in LLVM or JVM verification will raise
   an error.
+* Add `"enum"` `setup value`s representing enums in MIR verification.
 * The `SAW/create ghost variable` command and the associated
   `ghost variable value` value are now supported with JVM and MIR verification.